Moved Controllers and Views into their own folders
