+++ app/src/main/java/xterminators/spellingbee/App.java
@@ -8,9 +8,9 @@ import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.Arrays;
 
-import xterminators.spellingbee.cli.CLIController;
-import xterminators.spellingbee.cli.CLIView;
-import xterminators.spellingbee.gui.GuiView;
+import xterminators.spellingbee.controller.CLIController;
+import xterminators.spellingbee.view.CLIView;
+import xterminators.spellingbee.view.GuiView;
 
 public class App {
     public static void main(String[] args) {
