+++ app/src/test/java/xterminators/spellingbee/cli/CLIViewTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
 import xterminators.spellingbee.model.Rank;
+import xterminators.spellingbee.view.CLIView;
 
 public class CLIViewTest {
     private CLIView view;
